+++ src/test/scala/com/rayrobdod/json/builder/BuilderTest.scala
@@ -30,9 +30,7 @@ import org.scalatest.FunSpec;
 import scala.collection.immutable.Seq;
 import com.rayrobdod.json.builder.PiecewiseBuilder.Failures.{ExpectedPrimitive, UnsuccessfulTypeCoercion}
 import com.rayrobdod.json.testing.EnforcedFailure
-import com.rayrobdod.json.parser.Parser
-import com.rayrobdod.json.parser.FailureParser
-import com.rayrobdod.json.parser.IdentityParser
+import com.rayrobdod.json.parser.{Parser, FailureParser, IdentityParser, PrimitiveSeqParser}
 import com.rayrobdod.json.union.ParserRetVal
 import com.rayrobdod.json.union.ParserRetVal.{BuilderFailure, Complex, ParserFailure}
 import com.rayrobdod.json.testing.HexArrayStringConverter
@@ -50,65 +48,100 @@ class BuilderTest extends FunSpec {
 	
 	
 	describe("Builder.mapKey") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.mapKey[Int]{x => x}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("key", "value")){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.mapKey[Int]{case x => "key"}
-				builder.apply(builder.init, 1, "value",  new IdentityParser[String], ())
-			}
-		}
-		it ("passes through a parser's failure") {
-			assertResult(ParserFailure(EnforcedFailure)){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.mapKey[Int]{case x => "key"}
-				builder.apply(builder.init, 1, "value",  new FailureParser(EnforcedFailure), ())
-			}
-		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(EnforcedFailure, ())){
+		describe ("init") {
+			it ("is the same as Builder.init") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.mapKey[Int]{case x => "key"}
-				builder.apply(builder.init, 1, "value",  new IdentityParser[String], ())
+				assertResult(builder.init){builder.mapKey[Int]{x => x}.init}
+			}
+		}
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex(("key", "value"))){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapKey[Int]{x => "key"}
+					builder.apply(builder.init, 1, "value", new IdentityParser[String], "extra")
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapKey[Int]{x => "key"}
+					builder.apply(builder.init, 1, "value", new FailureParser(EnforcedFailure), "extra")
+				}
+			}
+			it ("passes through a builder's failure") {
+				assertResult(BuilderFailure(EnforcedFailure, "extra")){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.mapKey[Int]{x => "key"}
+					builder.apply(builder.init, 1, "value", new IdentityParser[String], "extra")
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
+				val builder = new ReportKeyValueBuilder[String, String]
+						.mapKey[Int]{x => "key"}
+				val middle = builder.apply(builder.init, 1, "value", new IdentityParser[String], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+						.mapKey[Int]{x => "key"}
+				assertResult( BuilderFailure(EnforcedFailure, "extra") ){ builder.finish("extra")(builder.init) }
 			}
 		}
 	}
 	describe("Builder.flatMapKey") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.flatMapKey[Int, Nothing]{case x => Right(x)}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("key", "value")){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapKey[Int, Nothing]{case x => Right("key")}
-				builder.apply(builder.init, 1, "value",  new IdentityParser[String], ())
-			}
-		}
-		it ("passes through a parser's failure") {
-			assertResult(ParserFailure(EnforcedFailure)){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapKey[Int, Nothing]{case x => Right("key")}
-				builder.apply(builder.init, 1, "value",  new FailureParser(EnforcedFailure), ())
-			}
-		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+		describe ("init") {
+			it ("is the same as Builder.init") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.flatMapKey[Int, Nothing]{case x => Right("key")}
-				builder.apply(builder.init, 1, "value",  new IdentityParser[String], ())
+				assertResult(builder.init){builder.flatMapKey[Int, Nothing]{x => Right(x)}.init}
 			}
 		}
-		it ("passes through a fun's failure") {
-			assertResult(BuilderFailure(Left(EnforcedFailure), ())){
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapKey[Int, Nothing]{x => Right("key")}
+					builder.apply(builder.init, 1, "value", new IdentityParser[String], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapKey[Int, Nothing]{x => Right("key")}
+					builder.apply(builder.init, 1, "value", new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("passes through a builder's failure") {
+				assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.flatMapKey[Int, Nothing]{x => Right("key")}
+					builder.apply(builder.init, 1, "value", new IdentityParser[String], ())
+				}
+			}
+			it ("passes through a fun's failure") {
+				assertResult(BuilderFailure(Left(EnforcedFailure), ())){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapKey[Int, EnforcedFailure.type]{x => Left(EnforcedFailure)}
+					builder.apply(builder.init, 1, "value", new IdentityParser[String], ())
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
 				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapKey[Int, EnforcedFailure.type]{case x => Left(EnforcedFailure)}
-				builder.apply(builder.init, 1, "value",  new IdentityParser[String], ())
+						.flatMapKey[Int, Nothing]{x => Right("key")}
+				val middle = builder.apply(builder.init, 1, "value", new IdentityParser[String], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+						.flatMapKey[Int, Nothing]{x => Right("key")}
+				assertResult( BuilderFailure(Right(EnforcedFailure), "extra") ){ builder.finish("extra")(builder.init) }
 			}
 		}
+		
 		
 		it ("when an inner builder fails, get the inner parser's extra") {
 			import com.rayrobdod.json.parser._
@@ -124,63 +157,97 @@ class BuilderTest extends FunSpec {
 		}
 	}
 	describe("Builder.mapValue") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.mapValue[String]{case x => x}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("key", "value")){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.mapValue[Int]{case x => "value"}
-				builder.apply(builder.init, "key", 1,  new IdentityParser[Int], ())
-			}
-		}
-		it ("passes through a parser's failure") {
-			assertResult(ParserFailure(EnforcedFailure)){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.mapValue[Int]{case x => "value"}
-				builder.apply(builder.init, "key", 1,  new FailureParser(EnforcedFailure), ())
-			}
-		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(EnforcedFailure, ())){
+		describe ("init") {
+			it ("is the same as Builder.init") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.mapValue[Int]{case x => "value"}
-				builder.apply(builder.init, "key", 1,  new IdentityParser[Int], ())
+				assertResult(builder.init){builder.mapValue[String]{x => x}.init}
+			}
+		}
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapValue[Int]{x => "value"}
+					builder.apply(builder.init, "key", 1, new IdentityParser[Int], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapValue[Int]{x => "value"}
+					builder.apply(builder.init, "key", 1, new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("passes through a builder's failure") {
+				assertResult(BuilderFailure(EnforcedFailure, ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.mapValue[Int]{x => "value"}
+					builder.apply(builder.init, "key", 1, new IdentityParser[Int], ())
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
+				val builder = new ReportKeyValueBuilder[String, String]
+						.mapValue[Int]{x => "value"}
+				val middle = builder.apply(builder.init, "key", 1, new IdentityParser[Int], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+						.mapValue[Int]{x => "key"}
+				assertResult( BuilderFailure(EnforcedFailure, "extra") ){ builder.finish("extra")(builder.init) }
 			}
 		}
 	}
 	describe("Builder.flatMapValue") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.flatMapValue[String, EnforcedFailure.type]{case x => Right(x)}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("key", "value")){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapValue[Int, EnforcedFailure.type]{case x => Right("value")}
-				builder.apply(builder.init, "key", 1,  new IdentityParser[Int], ())
-			}
-		}
-		it ("passes through a parser's failure") {
-			assertResult(ParserFailure(EnforcedFailure)){
-				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapValue[Int, EnforcedFailure.type]{case x => Right("value")}
-				builder.apply(builder.init, "key", 1,  new FailureParser(EnforcedFailure), ())
-			}
-		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+		describe ("init") {
+			it ("is the same as Builder.init") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.flatMapValue[Int, EnforcedFailure.type]{case x => Right("value")}
-				builder.apply(builder.init, "key", 1,  new IdentityParser[Int], ())
+				assertResult(builder.init){builder.flatMapValue[Int, Nothing]{x => Right(x)}.init}
 			}
 		}
-		it ("passes through a fun's failure") {
-			assertResult(BuilderFailure(Left(EnforcedFailure), ())){
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapValue[Int, Nothing]{x => Right("value")}
+					builder.apply(builder.init, "key", 1, new IdentityParser[Int], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapValue[Int, Nothing]{x => Right("value")}
+					builder.apply(builder.init, "key", 1, new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("passes through a builder's failure") {
+				assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.flatMapValue[Int, Nothing]{x => Right("value")}
+					builder.apply(builder.init, "key", 1, new IdentityParser[Int], ())
+				}
+			}
+			it ("passes through a fun's failure") {
+				assertResult(BuilderFailure(Left(EnforcedFailure), ())){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapValue[Int, EnforcedFailure.type]{x => Left(EnforcedFailure)}
+					builder.apply(builder.init, "key", 1, new IdentityParser[Int], ())
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
 				val builder = new ReportKeyValueBuilder[String, String]
-						.flatMapValue[Int, EnforcedFailure.type]{case x => Left(EnforcedFailure)}
-				builder.apply(builder.init, "key", 1,  new IdentityParser[Int], ())
+						.flatMapValue[Int, Nothing]{x => Right("value")}
+				val middle = builder.apply(builder.init, "key", 1, new IdentityParser[Int], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+						.flatMapValue[Int, Nothing]{x => Right("value")}
+				assertResult( BuilderFailure(Right(EnforcedFailure), "extra") ){ builder.finish("extra")(builder.init) }
 			}
 		}
 		
@@ -208,52 +275,151 @@ class BuilderTest extends FunSpec {
 			}
 		}
 	}
-	describe("Builder.mapResult") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.flatMapResult[String, EnforcedFailure.type]{case x => Right(x)}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("result")){
-				val builder = new ReportKeyValueBuilder[String, Int]
-						.mapResult[String]{x => "result"}
-				builder.finish(())(builder.init)
+	describe("Builder.mapFailure") {
+		describe ("init") {
+			it ("is the same as Builder.init") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+				assertResult(builder.init){builder.mapFailure[String]{x => "Failure"}.init}
 			}
 		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(EnforcedFailure, ())){
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapFailure{x => "Failure"}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapFailure{x => "Failure"}
+					builder.apply(builder.init, "key", "value", new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("maps a builder's failure") {
+				assertResult(BuilderFailure("Failure", ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.mapFailure{x => "Failure"}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
+				val builder = new ReportKeyValueBuilder[String, String]
+							.mapFailure{x => "Failure"}
+				val middle = builder.apply(builder.init, "key", "value", new IdentityParser[String], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.mapResult[String]{x => "result"}
-				builder.finish(())(builder.init)
+							.mapFailure{x => "Failure"}
+				assertResult( BuilderFailure("Failure", "extra") ){ builder.finish("extra")(builder.init) }
+			}
+		}
+	}
+	describe("Builder.mapResult") {
+		describe ("init") {
+			it ("is the same as Builder.init") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+				assertResult(builder.init){builder.mapResult{x => (x, "postmap")}.init}
+			}
+		}
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex(("key", "value"))){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapResult{x => (x._1, x._2, "postmap")}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.mapResult{x => (x._1, x._2, "postmap")}
+					builder.apply(builder.init, "key", "value", new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("maps a builder's failure") {
+				assertResult(BuilderFailure(EnforcedFailure, ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.mapResult{x:(String, String) => (x._1, x._2, "postmap")}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
+				val builder = new ReportKeyValueBuilder[String, String]
+							.mapResult{x => (x._1, x._2, "postmap")}
+				val middle = builder.apply(builder.init, "key", "value", new IdentityParser[String], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value", "postmap")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+							.mapResult{x:(String, String) => (x._1, x._2, "postmap")}
+				assertResult( BuilderFailure(EnforcedFailure, "extra") ){ builder.finish("extra")(builder.init) }
 			}
 		}
 	}
 	describe("Builder.flatMapResult") {
-		it ("The new builder's init is the same as the old builder's init") {
-			val builder = new ThrowBuilder(EnforcedFailure)
-			assertResult(builder.init){builder.flatMapResult[String, EnforcedFailure.type]{case x => Right(x)}.init}
-		}
-		it ("when a success, continue being a success") {
-			assertResult(Complex("result")){
-				val builder = new ReportKeyValueBuilder[String, Int]
-						.flatMapResult[String, EnforcedFailure.type]{case x => Right("result")}
-				builder.finish(())(builder.init)
-			}
-		}
-		it ("passes through a builder's failure") {
-			assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+		describe ("init") {
+			it ("is the same as Builder.init") {
 				val builder = new ThrowBuilder(EnforcedFailure)
-						.flatMapResult[String, EnforcedFailure.type]{case x => Right("result")}
-				builder.finish(())(builder.init)
+				assertResult(builder.init){builder.flatMapResult{x => Right((x, "postmap"))}.init}
 			}
 		}
-		it ("passes through a fun's failure") {
-			assertResult(BuilderFailure(Left(EnforcedFailure), ())){
-				val builder = new ReportKeyValueBuilder[String, Int]
-						.flatMapResult[String, EnforcedFailure.type]{case x => Left(EnforcedFailure)}
-				builder.finish(())(builder.init)
+		describe ("apply") {
+			it ("when a success, continue being a success") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapResult{x => Right((x._1, x._2, "postmap"))}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+			it ("passes through a parser's failure") {
+				assertResult(ParserFailure(EnforcedFailure)){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapResult{x => Right((x._1, x._2, "postmap"))}
+					builder.apply(builder.init, "key", "value", new FailureParser(EnforcedFailure), ())
+				}
+			}
+			it ("passes through a builder's failure") {
+				assertResult(BuilderFailure(Right(EnforcedFailure), ())){
+					val builder = new ThrowBuilder(EnforcedFailure)
+							.flatMapResult{x:(String, String) => Right((x._1, x._2, "postmap"))}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
+			}
+			it ("a fun's failure is irrelevant") {
+				assertResult(Complex("key", "value")){
+					val builder = new ReportKeyValueBuilder[String, String]
+							.flatMapResult{x => Left(EnforcedFailure)}
+					builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+				}
 			}
 		}
+		describe ("finish") {
+			it ("is a success if Builder.finish would have been a success") {
+				val builder = new ReportKeyValueBuilder[String, String]
+						.flatMapResult{x => Right((x._1, x._2, "postmap"))}
+				val middle = builder.apply(builder.init, "key", "value", new IdentityParser[String], "extra").fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult( Complex(("key", "value", "postmap")) ){ builder.finish("extra")(middle) }
+			}
+			it ("is a failure if Builder.finish would have been a failure") {
+				val builder = new ThrowBuilder(EnforcedFailure)
+						.flatMapResult{x:(String, String) => Right((x._1, x._2, "postmap"))}
+				assertResult( BuilderFailure(Right(EnforcedFailure), "extra") ){ builder.finish("extra")(builder.init) }
+			}
+			it ("passes through a fun's failure") {
+				val builder = new ReportKeyValueBuilder[String, String]
+						.flatMapResult{x => Left(EnforcedFailure)}
+				val middle = builder.apply(builder.init, "key", "value", new IdentityParser[String], ()).fold({x => x}, {x => ???}, {x => ???}, {(x,y) => ???})
+				assertResult(BuilderFailure(Left(EnforcedFailure), "extra")){ builder.finish("extra")(middle) }
+			}
+		}
+		
 		
 		it ("when an inner builder fails, get the inner parser's extra") {
 			import com.rayrobdod.json.parser._
@@ -269,26 +435,92 @@ class BuilderTest extends FunSpec {
 		}
 	}
 	describe("Builder.zip") {
-		it ("when a success, continue being a success (primitive)") {
-			val exp = Complex((("key", "value"), ("key", "value")))
-			val builder = new ReportKeyValueBuilder[String, String]
-					.zip(new ReportKeyValueBuilder[String, String])
-			
-			assertResult(exp){
-				builder.apply(builder.init, "key", "value", new IdentityParser[String], ())
+		import com.rayrobdod.json.parser._
+		import com.rayrobdod.json.union._
+		val cborBuilder = new CborBuilder()
+				.mapKey[StringOrInt].mapValue[JsonValue]
+		val jsonBuilder = new PrettyJsonBuilder(PrettyJsonBuilder.MinifiedPrettyParams)
+		val failureBuilder = new ThrowBuilder(UnsuccessfulTypeCoercion)
+				.mapKey[StringOrInt].mapValue[JsonValue]
+				.mapFailure[PiecewiseBuilder.Failures]
+		val primitiveBuilder = new ReportKeyValueBuilder[StringOrInt, JsonValue]
+				.mapFailure[PiecewiseBuilder.Failures]
+		
+		describe ("init") {
+			it ("asdf") {
+				val exp = cborBuilder.init -> jsonBuilder.init
+				val res = (cborBuilder zip jsonBuilder).init
+				assertResult(exp){res}
 			}
 		}
-		it ("when a success, continue being a success (complex)") {
-			import com.rayrobdod.json.parser._
-			import com.rayrobdod.json.union._
-			val failureMapping = {x:IllegalFoldingFailure.type => com.rayrobdod.json.builder.PrettyJsonBuilder.Failures.ArrayKeyNotIncrementing(0, 0)}
-			val exp = ParserRetVal.Complex( ((hexSeq"818163616263" ,"""[["abc"]]""")) )
-			val builder = new CborBuilder().mapKey[StringOrInt].mapValue[JsonValue].mapFailure(failureMapping)
-				.zip(new PrettyJsonBuilder(PrettyJsonBuilder.MinifiedPrettyParams))
+		describe ("apply") {
+			it ("success zip success -> success (primitive)") {
+				val exp = Complex(((StringOrInt("key"), JsonValue("value")), (StringOrInt("key"), JsonValue("value"))))
+				val zipBuilder = primitiveBuilder zip primitiveBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", JsonValue("value"), new IdentityParser[JsonValue], ())
+				assertResult(exp){res}
+			}
+			it ("failure zip success -> failure (primitive)") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = failureBuilder zip primitiveBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", JsonValue("value"), new IdentityParser[JsonValue], "extra")
+				assertResult(exp){res}
+			}
+			it ("success zip failure -> failure (primitive)") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = primitiveBuilder zip failureBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", JsonValue("value"), new IdentityParser[JsonValue], "extra")
+				assertResult(exp){res}
+			}
 			
-			val res = new JsonParser().parse(builder, """[ [ "abc" ] ]""")
-			assertResult(exp){res}
+			it ("success zip success -> success (complex)") {
+				val exp = {
+					val thing = CborBuilder.Middle(true,1,List(hexSeq"636b6579826476616c316476616c32"))
+					Complex((thing, thing))
+				}
+				val zipBuilder = cborBuilder zip cborBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", Seq("val1", "val2"), PrimitiveSeqParser[StringOrInt, JsonValue, String], ())
+				assertResult(exp){res}
+			}
+			it ("failure zip success -> failure (complex)") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = failureBuilder zip cborBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", Seq("val1", "val2"), PrimitiveSeqParser[StringOrInt, JsonValue, String], "extra")
+				assertResult(exp){res}
+			}
+			it ("success zip failure -> failure (complex)") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = cborBuilder zip failureBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", Seq("val1", "val2"), PrimitiveSeqParser[StringOrInt, JsonValue, String], "extra")
+				assertResult(exp){res}
+			}
+			
+			it ("passes through a parser's failure") {
+				val exp = ParserFailure(EnforcedFailure)
+				val zipBuilder = primitiveBuilder zip primitiveBuilder
+				val res = zipBuilder.apply(zipBuilder.init, "key", JsonValue("value"), new FailureParser(EnforcedFailure), "extra")
+				assertResult(exp){res}
+			}
+		}
+		describe ("finish") {
+			it ("success zip success -> success") {
+				val exp = Complex((hexSeq"80", "[]"))
+				val zipBuilder = cborBuilder zip jsonBuilder
+				val res = zipBuilder.finish("extra")(zipBuilder.init)
+				assertResult(exp){res}
+			}
+			it ("failure zip success -> failure") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = failureBuilder zip cborBuilder
+				val res = zipBuilder.finish("extra")(zipBuilder.init)
+				assertResult(exp){res}
+			}
+			it ("success zip failure -> failure") {
+				val exp = BuilderFailure(UnsuccessfulTypeCoercion, "extra")
+				val zipBuilder = cborBuilder zip failureBuilder
+				val res = zipBuilder.finish("extra")(zipBuilder.init)
+				assertResult(exp){res}
+			}
 		}
 	}
-	
 }