+++ CHANGES.md
@@ -2,6 +2,7 @@
 * Represent failures within the type system
   * Add a type parameter to Parser to represent ways that the Parser can fail
   * Add a type parameter to Builder to represent ways that the Builder can fail
+  * Add a type parameter to Parser to represent parser-related information about Builder failures
   * Replace `Failure` with `ParserFailure` and `BuilderFailure`
   * Add two type parameters to ParserRetVal (one for ParserFailure and one for BuilderFailure)
   * Make methods that used to return `Either[(String,Int),A]` return `ParserRetVal[A,Nothing,PF,BF]`